Windows support (#112)
* Windows support
