+++ libraries/platform_helpers.rb
@@ -1,0 +1,9 @@
+class Chef::Node
+  def windows?
+    self['os'] == 'windows'
+  end
+
+  def linux?
+    self['os'] == 'linux'
+  end
+end
