+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'peter@realityforge.org'
 license 'Apache-2.0'
 description 'Installs/Configures GlassFish Application Server'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '0.8.0'
+version '1.0.0'
 
 chef_version '>= 13.0' if respond_to?(:chef_version)
 
@@ -13,6 +13,7 @@ source_url 'https://github.com/realityforge/chef-glassfish'
 
 supports 'ubuntu'
 supports 'debian'
+supports 'windows'
 
 depends 'java'
 depends 'authbind'
