+++ providers/admin_object.rb
@@ -22,18 +22,21 @@ action :create do
   args << '--raname' << new_resource.raname
   args << '--restype' << new_resource.restype
   args << '--property' << encode_parameters(new_resource.properties) unless new_resource.properties.empty?
-  args << '--description' << "'#{new_resource.description}'" if new_resource.description
+  args << '--description' << "\"#{new_resource.description}\"" if new_resource.description
   args << '--classname' << new_resource.classname if new_resource.classname
   args << "--enabled=#{new_resource.enabled}" if new_resource.enabled
   args << asadmin_target_flag
   args << new_resource.name
 
   execute "asadmin_create-admin-object #{new_resource.name}" do
-    not_if "#{asadmin_command('list-admin-objects')} #{new_resource.target} | grep -F -x -- '#{new_resource.name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+
+    filter = pipe_filter(new_resource.name, regexp: false, line: true)
+    not_if "#{asadmin_command('list-admin-objects')} #{new_resource.target} | #{filter}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 end
 
@@ -44,10 +47,12 @@ action :delete do
   args << new_resource.name
 
   execute "asadmin_delete-admin-object #{new_resource.name}" do
-    only_if "#{asadmin_command('list-admin-objects')} #{new_resource.target} | grep -F -x -- '#{new_resource.name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+
+    only_if "#{asadmin_command('list-admin-objects')} #{new_resource.target} | #{pipe_filter(new_resource.name, regexp: false, line: true)}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 end
