+++ providers/auth_realm.rb
@@ -28,11 +28,11 @@ action :create do
   args << new_resource.name
 
   execute "asadmin_create_auth_realm #{new_resource.name}" do
-    not_if "#{asadmin_command('list-auth-realms')} #{new_resource.target} | grep -F -x -- '#{new_resource.name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+    not_if "#{asadmin_command('list-auth-realms')} #{new_resource.target} | #{pipe_filter(new_resource.name, regexp: false, line: true)}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 
   properties.each_pair do |key, value|
@@ -56,10 +56,11 @@ action :delete do
   args << new_resource.name
 
   execute "asadmin_delete_auth_realm #{new_resource.name}" do
-    only_if "#{asadmin_command('list-auth-realms')} #{new_resource.target} | grep -F -x -- '#{new_resource.name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+    filter = pipe_filter(new_resource.name, regexp: false, line: true)
+    only_if "#{asadmin_command('list-auth-realms')} #{new_resource.target} | #{filter}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 end
