+++ providers/context_service.rb
@@ -28,11 +28,12 @@ action :create do
   args << new_resource.jndi_name
 
   execute "asadmin_create-context-service #{new_resource.jndi_name}" do
-    not_if "#{asadmin_command('list-context-services')} #{new_resource.target} | grep -F -x -- '#{new_resource.jndi_name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+    filter = pipe_filter(new_resource.jndi_name, regexp: false, line: true)
+    not_if "#{asadmin_command('list-context-services')} #{new_resource.target} | #{filter}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 
   properties = {
@@ -64,10 +65,11 @@ action :delete do
   args << new_resource.jndi_name
 
   execute "asadmin_delete-context-service #{new_resource.jndi_name}" do
-    only_if "#{asadmin_command('list-context-services')} #{new_resource.target} | grep -F -x -- '#{new_resource.jndi_name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+    filter = pipe_filter(new_resource.jndi_name, regexp: false, line: true)
+    only_if "#{asadmin_command('list-context-services')} #{new_resource.target} | #{filter}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 end
