+++ providers/javamail_resource.rb
@@ -27,18 +27,19 @@ action :create do
   args << '--transprotocol' << new_resource.transprotocol if new_resource.transprotocol
   args << '--transprotocolclass' << new_resource.transprotocolclass if new_resource.transprotocolclass
   args << '--property' << encode_parameters(new_resource.properties) unless new_resource.properties.empty?
-  args << '--description' << "'#{new_resource.description}'" if new_resource.description
+  args << '--description' << "\"#{new_resource.description}\"" if new_resource.description
   args << "--debug=#{new_resource.debug}" if new_resource.debug
   args << "--enabled=#{new_resource.enabled}" if new_resource.enabled
   args << asadmin_target_flag
   args << new_resource.jndi_name
 
   execute "asadmin_create-javamail-resource #{new_resource.jndi_name}" do
-    not_if "#{asadmin_command('list-javamail-resources')} #{new_resource.target} | grep -F -x -- '#{new_resource.jndi_name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+    filter = pipe_filter(new_resource.jndi_name, regexp: false, line: true)
+    not_if "#{asadmin_command('list-javamail-resources')} #{new_resource.target} | #{filter}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 end
 
@@ -49,10 +50,11 @@ action :delete do
   args << new_resource.jndi_name
 
   execute "asadmin_delete-javamail-resource #{new_resource.jndi_name}" do
-    only_if "#{asadmin_command('list-javamail-resources')} #{new_resource.target} | grep -F -x -- '#{new_resource.jndi_name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+    filter = pipe_filter(new_resource.jndi_name, regexp: false, line: true)
+    only_if "#{asadmin_command('list-javamail-resources')} #{new_resource.target} | #{filter}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 end
