+++ providers/jvm_options.rb
@@ -16,13 +16,28 @@
 
 include Chef::Asadmin
 
+def service_name
+  "glassfish-#{new_resource.domain_name}"
+end
+
 action :set do
-  service "glassfish-#{new_resource.domain_name}" do
-    supports restart: true, status: true
+  glassfish_wait_for_glassfish new_resource.domain_name do
+    username new_resource.username
+    password_file new_resource.password_file
+    admin_port new_resource.admin_port
+    only_if { new_resource.admin_port }
     action :nothing
   end
 
-  output = `#{asadmin_command('list-jvm-options', true, terse: true, echo: false)}`
+  service service_name do
+    supports restart: true, status: true
+    action :nothing
+    notifies :run, "glassfish_wait_for_glassfish[#{new_resource.domain_name}]", :immediately
+  end
+
+  require 'mixlib/shellout'
+
+  output = Mixlib::ShellOut.new(asadmin_command('list-jvm-options', true, terse: true, echo: false)).run_command.stdout
 
   # Work around bugs in 3.1.2.2
   if node['glassfish']['version'] == '3.1.2.2'
@@ -36,8 +51,8 @@ action :set do
 
       execute "asadmin_create-jvm-option #{line}" do
         timeout node['glassfish']['asadmin']['timeout'] + 5
-        user new_resource.system_user unless node['os'] == 'windows'
-        group new_resource.system_group unless node['os'] == 'windows'
+        user new_resource.system_user unless node.windows?
+        group new_resource.system_group unless node.windows?
         command asadmin_command(args.join(' '))
 
         notifies :restart, "service[glassfish-#{new_resource.domain_name}]", :delayed
@@ -52,37 +67,50 @@ action :set do
 
       execute "asadmin_delete-jvm-option #{line}" do
         timeout node['glassfish']['asadmin']['timeout'] + 5
-        user new_resource.system_user unless node['os'] == 'windows'
-        group new_resource.system_group unless node['os'] == 'windows'
+        user new_resource.system_user unless node.windows?
+        group new_resource.system_group unless node.windows?
         command asadmin_command(args.join(' '))
 
         notifies :restart, "service[glassfish-#{new_resource.domain_name}]", :delayed
       end
     end
   else
-    existing = output.split("\n")
+    existing = output.split("\n") if node.linux?
+    existing = output.split("\r\n") if node.windows?
 
     existing_option_string = encode_options(existing)
     new_option_string = encode_options(new_resource.options)
 
     if existing_option_string != new_option_string
-      delete_command = []
-      delete_command << 'delete-jvm-options'
-      delete_command << existing_option_string
-      delete_command << asadmin_target_flag
+      execute "asadmin_delete-jvm-options #{new_resource.name}" do
+        delete_command = []
+        delete_command << 'delete-jvm-options'
+        delete_command << existing_option_string
+        delete_command << asadmin_target_flag
 
-      create_command = []
-      create_command << 'create-jvm-options'
-      create_command << new_option_string
-      create_command << asadmin_target_flag
-
-      execute "asadmin_set-jvm-options #{new_resource.name}" do
         timeout node['glassfish']['asadmin']['timeout'] + 5
-        user new_resource.system_user unless node['os'] == 'windows'
-        group new_resource.system_group unless node['os'] == 'windows'
-        command "#{asadmin_command(delete_command.join(' '))} && #{asadmin_command(create_command.join(' '))}"
 
-        notifies :restart, "service[glassfish-#{new_resource.domain_name}]", :immediate
+        user new_resource.system_user unless node.windows?
+        group new_resource.system_group unless node.windows?
+        command asadmin_command(delete_command.join(' '))
+
+        notifies :run, "execute[asadmin_create-jvm-options #{new_resource.name}]", :immediate
+      end
+
+      execute "asadmin_create-jvm-options #{new_resource.name}" do
+        create_command = []
+        create_command << 'create-jvm-options'
+        create_command << new_option_string
+        create_command << asadmin_target_flag
+
+        timeout node['glassfish']['asadmin']['timeout'] + 5
+
+        user new_resource.system_user unless node.windows?
+        group new_resource.system_group unless node.windows?
+        command asadmin_command(create_command.join(' '))
+
+        action :nothing
+        notifies :restart, "service[#{service_name}]", :immediate
       end
     end
   end
