+++ providers/managed_executor_service.rb
@@ -36,11 +36,12 @@ action :create do
   args << new_resource.jndi_name
 
   execute "asadmin_create-managed-executor-service #{new_resource.jndi_name}" do
-    not_if "#{asadmin_command('list-managed-executor-services')} #{new_resource.target} | grep -F -x -- '#{new_resource.jndi_name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+    filter = pipe_filter(new_resource.jndi_name, regexp: false, line: true)
+    not_if "#{asadmin_command('list-managed-executor-services')} #{new_resource.target} | #{filter}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 
   properties = {
@@ -73,16 +74,17 @@ action :create do
 end
 
 action :delete do
-  command = []
-  command << 'delete-managed-executor-service'
-  command << asadmin_target_flag
-  command << new_resource.jndi_name
+  args = []
+  args << 'delete-managed-executor-service'
+  args << asadmin_target_flag
+  args << new_resource.jndi_name
 
   execute "asadmin_delete-managed-executor-service #{new_resource.jndi_name}" do
-    only_if "#{asadmin_command('list-managed-executor-services')} #{new_resource.target} | grep -F -x -- '#{new_resource.jndi_name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
-    command asadmin_command(command.join(' '))
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
+    command asadmin_command(args.join(' '))
+    filter = pipe_filter(new_resource.jndi_name, regexp: false, line: true)
+    only_if "#{asadmin_command('list-managed-executor-services')} #{new_resource.target} | #{filter}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 end
