+++ providers/web_env_entry.rb
@@ -21,7 +21,7 @@ action :set do
   args << 'set-web-env-entry'
   args << '--name' << new_resource.name
   args << '--type' << new_resource.type
-  args << '--description' << "'#{new_resource.description}'" if new_resource.description
+  args << '--description' << "\"#{new_resource.description}\"" if new_resource.description
   args << if new_resource.value.nil?
             '--ignoreDescriptorItem'
           else
@@ -30,11 +30,12 @@ action :set do
   args << new_resource.webapp
 
   execute "asadmin_set-web-env-entry #{new_resource.webapp} --name #{new_resource.name}" do
-    not_if "#{asadmin_command("list-web-env-entry #{new_resource.webapp}")} | grep -F -x -- '#{new_resource.name} (#{new_resource.type}) #{new_resource.value} ignoreDescriptorItem=#{new_resource.value.nil?} //(#{new_resource.description || 'description not specified'})'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+    filter = pipe_filter("#{new_resource.name} (#{new_resource.type}) #{new_resource.value} ignoreDescriptorItem=#{new_resource.value.nil?} //(#{new_resource.description || 'description not specified'})", regexp: false, line: true)
+    not_if "#{asadmin_command("list-web-env-entry #{new_resource.webapp}")} | #{filter}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 end
 
@@ -45,10 +46,11 @@ action :unset do
   args << new_resource.webapp
 
   execute "asadmin_unset-web-env-entry #{new_resource.name}" do
-    only_if "#{asadmin_command("list-web-env-entry #{new_resource.webapp}")} | grep -F -x -- '#{new_resource.name}'", timeout: node['glassfish']['asadmin']['timeout'] + 5
     timeout node['glassfish']['asadmin']['timeout'] + 5
-    user new_resource.system_user unless node['os'] == 'windows'
-    group new_resource.system_group unless node['os'] == 'windows'
+    user new_resource.system_user unless node.windows?
+    group new_resource.system_group unless node.windows?
     command asadmin_command(args.join(' '))
+    filter = pipe_filter(new_resource.name, regexp: false, line: true)
+    only_if "#{asadmin_command("list-web-env-entry #{new_resource.webapp}")} | #{filter}", timeout: node['glassfish']['asadmin']['timeout'] + 5
   end
 end
