+++ resources/domain.rb
@@ -85,6 +85,8 @@ attribute :secure, kind_of: [TrueClass, FalseClass], default: false
 attribute :logging_properties, kind_of: Hash, default: {}
 # <> @attribute realm_types A map of names to realm implementation classes that is merged into the default realm types.
 attribute :realm_types, kind_of: Hash, default: {}
+# <> @attribute certificate_cn The common name that should be used when generating the self-signed ssl certificate for the domain
+attribute :certificate_cn, kind_of: String, default: nil
 
 # <> @attribute system_user The user that the domain executes as. Defaults to `node['glassfish']['user']` if unset.
 attribute :system_user, kind_of: String, default: nil
@@ -135,15 +137,12 @@ end
 
 def installation_jvm_options
   [
-    # TODO: All jvm options that have expanded node['glassfish']['install_dir'] should be replaced by ${com.sun.aas.installRoot} in modern glassfish versions and should also use${path.separator}
     "-Dcom.sun.aas.instanceRoot=#{domain_dir_path}",
     '-Dcom.sun.enterprise.config.config_environment_factory_class=com.sun.enterprise.config.serverbeans.AppserverConfigEnvironmentFactory',
-    # TODO: Next line is not needed as of modern glassfish
-    "-Dcom.sun.aas.installRoot=#{node['glassfish']['install_dir']}/glassfish",
     '-DANTLR_USE_DIRECT_CLASS_LOADING=true',
-    "-javaagent:#{node['glassfish']['install_dir']}/glassfish/lib/monitor/flashlight-agent.jar",
-    "-Djava.ext.dirs=#{node['java']['java_home']}/lib/ext:#{node['java']['java_home']}/jre/lib/ext:#{domain_dir_path}/lib/ext",
-    "-Djava.endorsed.dirs=#{node['glassfish']['install_dir']}/glassfish/modules/endorsed:#{node['glassfish']['install_dir']}/glassfish/lib/endorsed",
+    '-javaagent:${com.sun.aas.installRoot}/lib/monitor/flashlight-agent.jar',
+    "-Djava.ext.dirs=#{node['java']['java_home']}/lib/ext${path.separator}#{node['java']['java_home']}/jre/lib/ext${path.separator}#{domain_dir_path}/lib/ext",
+    '-Djava.endorsed.dirs=${com.sun.aas.installRoot}/modules/endorsed${path.separator}${com.sun.aas.installRoot}/lib/endorsed',
   ]
 end
 
@@ -151,7 +150,7 @@ def osgi_jvm_options
   [
     '-Dosgi.shell.telnet.maxconn=1',
     '-Dfelix.fileinstall.disableConfigSave=false',
-    "-Dfelix.fileinstall.dir=#{node['glassfish']['install_dir']}/glassfish/modules/autostart/",
+    '-Dfelix.fileinstall.dir=${com.sun.aas.installRoot}/modules/autostart/',
     '-Dosgi.shell.telnet.port=6666',
     '-Dfelix.fileinstall.log.level=2',
     '-Dfelix.fileinstall.poll=5000',
@@ -192,4 +191,116 @@ def jvm_options
   default_jvm_options +
     java_agents.map { |agent| "-javaagent:#{agent}" } +
     extra_jvm_options
+end
+
+def default_logging_properties
+  {
+    'handlers' => 'java.util.logging.ConsoleHandler',
+    'handlerServices' => 'com.sun.enterprise.server.logging.GFFileHandler,com.sun.enterprise.server.logging.SyslogHandler',
+
+    'java.util.logging.ConsoleHandler.formatter' => 'com.sun.enterprise.server.logging.UniformLogFormatter',
+
+    'com.sun.enterprise.server.logging.GFFileHandler.formatter' => 'com.sun.enterprise.server.logging.UniformLogFormatter',
+    'com.sun.enterprise.server.logging.GFFileHandler.file' => '${com.sun.aas.instanceRoot}/logs/server.log',
+    'com.sun.enterprise.server.logging.GFFileHandler.rotationTimelimitInMinutes' => '0',
+    'com.sun.enterprise.server.logging.GFFileHandler.flushFrequency' => '1',
+    'com.sun.enterprise.server.logging.GFFileHandler.logtoConsole' => 'false',
+    'com.sun.enterprise.server.logging.GFFileHandler.rotationLimitInBytes' => '2000000',
+    'com.sun.enterprise.server.logging.GFFileHandler.retainErrorsStasticsForHours' => '0',
+    'com.sun.enterprise.server.logging.GFFileHandler.maxHistoryFiles' => '3',
+    'com.sun.enterprise.server.logging.GFFileHandler.rotationOnDateChange' => 'false',
+
+    'com.sun.enterprise.server.logging.SyslogHandler.useSystemLogging' => 'false',
+
+    'log4j.logger.org.hibernate.validator.util.Version' => 'warn',
+
+    # Payara 5.182
+    'fish.payara.enterprise.server.logging.PayaraNotificationFileHandler.compressOnRotation' => 'false',
+    'fish.payara.enterprise.server.logging.PayaraNotificationFileHandler.rotationLimitInBytes' => '2000000',
+    'fish.payara.enterprise.server.logging.PayaraNotificationFileHandler.rotationOnDateChange' => 'false',
+    'fish.payara.enterprise.server.logging.PayaraNotificationFileHandler.file' => '${com.sun.aas.instanceRoot}/logs/notification.log',
+    'fish.payara.enterprise.server.logging.PayaraNotificationFileHandler.logtoFile' => 'true',
+    'fish.payara.enterprise.server.logging.PayaraNotificationFileHandler.maxHistoryFiles' => '0',
+    'com.sun.enterprise.server.logging.GFFileHandler.logtoFile' => 'true',
+    'fish.payara.enterprise.server.logging.PayaraNotificationFileHandler.rotationTimelimitInMinutes' => '0',
+
+    # Payara 5.183
+    'fish.payara.enterprise.server.logging.PayaraNotificationFileHandler.formatter' => 'com.sun.enterprise.server.logging.ODLLogFormatter',
+
+    # All log level details
+    '.level' => 'INFO',
+
+    'com.sun.enterprise.server.logging.GFFileHandler.level' => 'ALL',
+    'javax.enterprise.system.tools.admin.level' => 'INFO',
+    'org.apache.jasper.level' => 'INFO',
+    'javax.enterprise.resource.corba.level' => 'INFO',
+    'javax.enterprise.system.core.level' => 'INFO',
+    'javax.enterprise.system.core.classloading.level' => 'INFO',
+    'java.util.logging.ConsoleHandler.level' => 'FINEST',
+    'javax.enterprise.system.webservices.saaj.level' => 'INFO',
+    'javax.enterprise.system.tools.deployment.level' => 'INFO',
+    'javax.enterprise.system.container.ejb.level' => 'INFO',
+    'javax.enterprise.system.core.transaction.level' => 'INFO',
+    'org.apache.catalina.level' => 'INFO',
+    'javax.enterprise.system.container.ejb.mdb.level' => 'INFO',
+    'org.apache.coyote.level' => 'INFO',
+    'javax.level' => 'INFO',
+    'javax.enterprise.resource.javamail.level' => 'INFO',
+    'javax.enterprise.system.webservices.rpc.level' => 'INFO',
+    'javax.enterprise.system.container.web.level' => 'INFO',
+    'javax.enterprise.system.util.level' => 'INFO',
+    'javax.enterprise.resource.resourceadapter.level' => 'INFO',
+    'javax.enterprise.resource.jms.level' => 'INFO',
+    'javax.enterprise.system.core.config.level' => 'INFO',
+    'javax.enterprise.system.level' => 'INFO',
+    'javax.enterprise.system.core.security.level' => 'INFO',
+    'javax.enterprise.system.container.cmp.level' => 'INFO',
+    'javax.enterprise.system.webservices.registry.level' => 'INFO',
+    'javax.enterprise.system.core.selfmanagement.level' => 'INFO',
+    'javax.enterprise.resource.jdo.level' => 'INFO',
+    'javax.enterprise.system.core.naming.level' => 'INFO',
+    'javax.enterprise.resource.webcontainer.jsf.application.level' => 'INFO',
+    'javax.enterprise.resource.webcontainer.jsf.resource.level' => 'INFO',
+    'javax.enterprise.resource.webcontainer.jsf.config.level' => 'INFO',
+    'javax.enterprise.resource.webcontainer.jsf.context.level' => 'INFO',
+    'javax.enterprise.resource.webcontainer.jsf.facelets.level' => 'INFO',
+    'javax.enterprise.resource.webcontainer.jsf.lifecycle.level' => 'INFO',
+    'javax.enterprise.resource.webcontainer.jsf.managedbean.level' => 'INFO',
+    'javax.enterprise.resource.webcontainer.jsf.renderkit.level' => 'INFO',
+    'javax.enterprise.resource.webcontainer.jsf.taglib.level' => 'INFO',
+    'javax.enterprise.resource.webcontainer.jsf.timing.level' => 'INFO',
+    'javax.enterprise.resource.jta.level' => 'WARNING',
+    'javax.enterprise.resource.sqltrace.level' => 'FINE',
+    'javax.org.glassfish.persistence.level' => 'INFO',
+    'org.jvnet.hk2.osgiadapter.level' => 'INFO',
+    'javax.enterprise.system.tools.backup.level' => 'INFO',
+    'org.glassfish.admingui.level' => 'INFO',
+    'javax.enterprise.system.ssl.security.level' => 'INFO',
+    'ShoalLogger.level' => 'CONFIG',
+    'org.eclipse.persistence.session.level' => 'INFO',
+    'javax.enterprise.resource.resourceadapter.com.sun.gjc.spi.level' => 'WARNING',
+    'com.hazelcast.level' => 'WARNING',
+  }
+end
+
+def default_realm_confs
+  common_confs = {
+    'fileRealm' => 'com.sun.enterprise.security.auth.login.FileLoginModule',
+    'ldapRealm' => 'com.sun.enterprise.security.auth.login.LDAPLoginModule',
+    'solarisRealm' => 'com.sun.enterprise.security.auth.login.SolarisLoginModule',
+  }
+
+  if node['glassfish']['version'][0].to_i >= 4
+    {
+      'jdbcRealm' => 'com.sun.enterprise.security.ee.auth.login.JDBCLoginModule',
+      'jdbcDigestRealm' => 'com.sun.enterprise.security.ee.auth.login.JDBCDigestLoginModule',
+      'pamRealm' => 'com.sun.enterprise.security.ee.auth.login.PamLoginModule',
+    }.merge common_confs
+  else
+    {
+      'jdbcRealm' => 'com.sun.enterprise.security.auth.login.JDBCLoginModule',
+      'jdbcDigestRealm' => 'com.sun.enterprise.security.auth.login.JDBCDigestLoginModule',
+      'pamRealm' => 'com.sun.enterprise.security.auth.login.PamLoginModule',
+    }.merge common_confs
+  end
 end
