+++ templates/default/asenv.conf.erb
@@ -27,3 +27,4 @@ AS_INSTALL="<%= "#{node['glassfish']['install_dir']}/glassfish" %>"
 AS_DEF_DOMAINS_PATH="<%= node['glassfish']['domains_dir'] %>"
 AS_DEF_NODES_PATH="../nodes"
 AS_DERBY_INSTALL="../../javadb"
+AS_H2_INSTALL="../../h2db"
