+++ templates/default/password.erb
@@ -1,2 +1,5 @@
-AS_ADMIN_MASTERPASSWORD=<%= @master_password %>
-AS_ADMIN_PASSWORD=<%= @password %>
+<% if @extend %><%= File.read(@extend) %><% end %>
+<% if @master_password %>AS_ADMIN_MASTERPASSWORD=<%= @master_password %><% end %>
+<% if @password %>AS_ADMIN_PASSWORD=<%= @password %><% end %>
+<% if @user_password %>AS_ADMIN_USERPASSWORD=<%= @user_password %><% end %>
+
